+++ .storybook/main.js
@@ -1,4 +1,8 @@
+const { setupPostcss } = require('../config-webpack-postcss');
 module.exports = {
+  webpackFinal: config => {
+    return setupPostcss(config);
+  },
   stories: ['../src/**/*.stories.js'],
   addons: [
     '@storybook/preset-create-react-app',
