+++ .storybook/preview-body.html
@@ -1,0 +1,1 @@
+<body class="font-body"></body>
